backend/requirements.txt
Django==5.0.3
psycopg[binary]
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-filter==24.2
django-cors-headers==4.3.1
django-csp==3.8
python-dotenv==1.0.1
Pillow>=11.0.0
whitenoise==6.6.0
dj-database-url==2.1.0
django-axes==6.4.0
django-otp==1.3.0
django-redis==5.4.0
redis==5.0.1
drf-spectacular==0.27.2
pytest>=8.3.3
pytest-django>=4.9.0
webauthn==2.5.1

